+++ syft/serde/protobuf/serde.py
@@ -13,6 +13,7 @@ from syft.generic.pointers.pointer_tensor import PointerTensor
 from syft.messaging.message import ObjectMessage
 from syft.messaging.message import TensorCommandMessage
 from syft.execution.plan import Plan
+from syft.execution.plan import NestedTypeWrapper
 from syft.execution.protocol import Protocol
 from syft.execution.role import Role
 from syft.execution.state import State
@@ -51,6 +52,7 @@ OBJ_PROTOBUF_TRANSLATORS = [
     PlaceholderId,
     PlaceHolder,
     Plan,
+    NestedTypeWrapper,
     PointerTensor,
     Protocol,
     Role,
